requirements.txt
# Python dependencies
# Install with: pip install -r requirements.txt

playwright
python-dotenv
pylance
